+++ ContentPatcher/Framework/Migrations/Migration_1_0.cs
@@ -1,0 +1,17 @@
+using System.Diagnostics.CodeAnalysis;
+using StardewModdingAPI;
+
+namespace ContentPatcher.Framework.Migrations
+{
+    /// <summary>Migrate patches to format version 1.3.</summary>
+    [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Named for clarity.")]
+    internal class Migration_1_0 : BaseMigration
+    {
+        /*********
+        ** Public methods
+        *********/
+        /// <summary>Construct an instance.</summary>
+        public Migration_1_0()
+            : base(new SemanticVersion(1, 0, 0)) { }
+    }
+}