+++ ContentPatcher/release-notes.md
@@ -6,6 +6,7 @@
 * Added `FromFile` patch token (e.g. so you can do `"HasFile:{{FromFile}}": true`).
 * Fixed `{{Range}}` not including inclusive upper bound.
 * Fixed validation for `Target` fields containing `{{Target}}` and `{{TargetWithoutPath}}` tokens.
+* Fixed validation for `Target` fields not shown in `patch summary` in some cases.
 * Fixed 'file does not exist' error when the `FromFile` path is ready and doesn't exist, but the patch is disabled by a patch-specific condition.
 * Fixed error when removing a map tile without edits.
 * Fixed token handling in map tile/property fields.